requirements.txt
Flask==2.2.3
SQLAlchemy==2.0.9
PyMySQL==1.0.3
redis==4.5.4
python-dotenv==1.0.0
Flask-JWT-Extended==4.2.3
werkzeug==2.2.3
cryptography==40.0.2